+++ packages/path_provider/path_provider_linux/lib/src/get_application_id_stub.dart
@@ -2,4 +2,5 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-export 'src/path_provider_linux.dart';
+/// Gets the application ID for this app.
+String? getApplicationId() => null;